Extract DrawSystem
